requirements.txt
absl-py==1.4.0
addict==2.4.0
aiofiles==22.1.0
aiosqlite==0.18.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asciitree==0.3.3
asttokens==2.2.1
astunparse==1.6.3
atomicwrites==1.4.1
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.2
black==22.3.0
bleach==6.0.0
boto3==1.34.3
botocore==1.34.3
Brotli==1.1.0
cachetools==5.3.3
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1700303426725/work/certifi
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.0.1
click==8.1.7
cloud-files==4.21.1
cloud-volume==8.28.0
cloudpickle==3.0.0
comm==0.1.2
compressed-segmentation==2.2.2
compresso==3.2.2
conda-pack @ file:///croot/conda-pack_1710258022278/work
contourpy==1.0.7
crackle-codec==0.7.1
crc32c==2.3.post0
cycler==0.11.0
Cython==0.29.21
daal==2024.0.1
daal4py==2024.0.1
debugpy==1.6.6
decorator==5.1.1
deflate==0.5.0
defusedxml==0.7.1
dill==0.3.7
dnspython==2.6.1
docker-pycreds==0.4.0
DracoPy==1.3.0
einops==0.6.0
exceptiongroup==1.2.0
executing==1.2.0
fairscale==0.4.13
fasteners==0.19
fastjsonschema==2.16.3
fastremap==1.14.1
filelock==3.15.4
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.38.0
fpzip==1.2.3
fqdn==1.5.1
fsspec==2024.6.1
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.4.0
gevent==23.9.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.15.0
google-apitools==0.5.32
google-auth==2.26.1
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
GPUtil==1.4.0
greenlet==3.0.3
grpcio==1.51.1
h5py==3.8.0
httplib2==0.22.0
huggingface-hub==0.24.5
hydra-core==1.3.2
idna==3.4
imageio==2.26.0
importlib-resources==5.12.0
importlib_metadata==8.2.0
inflection==0.5.1
iniconfig==2.0.0
iopath==0.1.9
ipykernel==6.21.2
ipympl==0.9.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.2
jupyter-events==0.6.3
jupyter-ydoc==0.2.2
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_fileid==0.8.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.19.0
keras==2.13.1
kiwisolver==1.4.4
lazy_loader==0.1
libclang==16.0.6
llvmlite==0.41.1
mahotas==1.4.13
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mmcv==2.2.0
mmengine==0.10.4
monai==1.1.0
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
neuroglancer==2.36
notebook==6.5.2
notebook_shim==0.2.2
numba==0.58.1
numcodecs==0.12.1
numpy==1.24.2
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
opt-einsum==3.3.0
orjson==3.9.10
packaging==23.0
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathos==0.3.1
pathspec==0.12.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
plotly==5.18.0
pluggy==1.3.0
portalocker==2.7.0
posix-ipc==1.1.1
pox==0.3.3
ppft==1.7.6.7
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.23.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.11.1
pycocotools==2.0.7
pycparser==2.21
pydot==3.0.1
Pygments==2.14.0
pymongo==4.8.0
pyparsing==3.0.9
pyrsistent==0.19.3
pysimdjson==5.0.2
pyspng-seunglab==1.1.1
pytest==7.4.4
python-dateutil==2.8.2
python-json-logger==2.0.7
python-jsonschema-objects==0.5.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.16.2
rsa==4.9
s3transfer==0.9.0
safetensors==0.4.3
scikit-image==0.20.0
scikit-learn==1.2.1
scipy==1.9.1
Send2Trash==1.8.0
sentry-sdk==2.8.0
setproctitle==1.3.3
simplejpeg==1.7.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
tabulate==0.9.0
tbb==2021.11.0
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.2.28
timm==1.0.8
tinycss2==1.2.1
tomli==2.0.1
torch==1.10.0+cu111
torchaudio==0.10.0+rocm4.1
torchvision==0.11.0+cu111
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.4
uri-template==1.2.0
urllib3==1.26.14
wandb==0.17.4
waterz==0.9.4
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wrapt==1.15.0
y-py==0.5.9
yacs==0.1.8
yapf==0.40.2
ypy-websocket==0.8.2
zarr==2.16.1
zfpc==0.1.2
zfpy==1.0.0
zipp==3.20.2
zope.event==5.0
zope.interface==6.1
zstandard==0.22.0
-v -e ./submodules/cremi_python